Use a more efficient UUID generation than UUID.randomUUID()
In perf analysis, UUID.randomUUID() has been shown to have a significant overhead in inserts
particularly on systems that may not have efficient random data collectors (e.g. daemons like
    haveged or hardware RNG). This new one combines the unique system-wide long for the region
and combines it with a locally generated pseudo-random long and the combo is always guaranteed
to be unique (we never expect the system-wide long to overflow).