Handling parameter constant of type string for prepare statement. (#157)
In case column is of "String" data type i.e. alias of Clob, when user set parameter value, that is internally created as SQLClob. But when this value is being extracted, being a Clob it gives its value as SQLClob@0xeeb6b6e.

To resolve same, if datatype of column is SQLClob, its char-array is taken and converted to string that works fine.