Turn conserve-sockets=false by default
With the default connection pooling, the scalability problem with messaging is largely
solved hence turning off conserve-sockets. Having selectors for receivers does not help
in terms of performance in modern Linux distributions using NPTL where there is no overhead
in having idle threads. The only advantage selector brings is lesser amount of memory with
lesser number of threads (which will only matter beyond thousands of nodes). But it also
brings a significant overhead of thread-context switching which can be alleviated by
having a hybrid server like the one implemented for Thrift.