Explicitly mark Connection as pooled
To avoid possible ambiguities with "sharedResource" and "unordered", now explicitly marking
a connection as pooled which is checked when trying to return to the pool.

Use blocking channel mode for non-pooled connections.

Skip the minimum pool size since background thread trying to create connections
(to meet minimum criteria) is causing trouble occasionally especially in startup/shutdown.