[SNAP-1410] #163
Changes proposed in this pull request

Execution of a prepared statement returns results in chunks of size DML_MAX_CHUNK_SIZE. So, SnappyDataServiceImpl.executePreparedQuery exits if the result of the size is greater than DML_MAX_CHUNK_SIZE. The remaining records are fetched using scroll cursor. But while exiting, SnappyDataServiceImpl.executePreparedQuery clears the parameters of PreparedExecution and hence scroll cursor returns either no result or incorrect results. This was seen while fetching the cached batches from smart connector mode.

To fix this, avoided cleaning of parameters if the batch is not the last batch. The parameter will anyway be cleared in the next execution of the prepared statement.
Patch testing

precheckin. Tested with the failing queries.