Snap 1304 new (#149)
For Design Please look at : https://docs.google.com/document/d/1bxwSTtk9mW6_qaoubtMsNwm7LGAzBrJA0yTENwIh5uw/edit#heading=h.2h7zw4ru3nw7

We have provided API's to begin/commit a transaction with snapshot isolation. With this, a TXState is created at the time of iteration and all the versions for write operations are recorded in the TXState. All the DistributionMessages now carry txId which is used to either create/use TXState on peer nodes.

For all the operations, we directly apply the changes to the underlying region but record the version in the TXState. The recorded version in TXState are recorded in snapshot RVV at the time of commit.

For all the write operations, a copy of old entry to an oldEntryMap at cache level is made. The oldEntryMap is used to store older entry so that if a query having older snapshot can find corresponding entries in oldEntryMap.
Old entries are cleaned by a thread at regular interval.

For read operations, compare the version of RegionEntry with the version stored in snapshot before returning them. If the version of RegionEntry is higher then the entry is fethced from the oldEntryMap maintained.