update UnsafeAtomic* impls with JDK8 Unsafe methods
Use the more efficient JDK8 Unsafe methods for getAndAdd/getAndSet in UnsafeAtomic* impls.

Update ConcurrentMapOpsTest to use the real scenario where field in base class but update
has to be done from a child class object.

As of latest JDK8 the difference between Unsafe* versions and JDK versions is only
~10% on a laptop (unlike with JDK6/7) so need to see if these classes should be dropped.