[SNAP-1651] Mark partition result owners to apply different eviction policy (#208)
Removed the array of nonEvictingOwners and will use a specific check for partition result on SnappyData side to indicate that the eviction policy change will apply only for partition result (because it applies a specific size limit for eviction and not zero eviction). Also the same will now be applied for the on-heap byte[] (and not just the off-heap buffer) created at the end because that too can fail with low memory error frequently.