Smart conn (#209)
It may happen that, zipped tasks in snappy will use different connection on which will be serviced by different nodes. In case of one task starting the snapshot tx, and other task using it but on different node, we can either commit it ignoring that it is not a co-ordinator or add to affected regions. (that we don't know when using through procedure)
In case of snapshot ignoring it so that txState created at each node using procedure can be committed individually.

Have disabled a logging which was causing NPE.