Fixing a deadlock seen in FabricServiceImpl
Deadlock between a thread obtaining a connection on a VM coming up or stopping:

TopService -> FabricServiceImpl sync in notifyRunning/Stop

and the regular FabricService start/stop will do the reverse:

FabricServiceImpl sync -> TopService (when creating EmbedConnection)

Now avoiding synchronized in notify* methods in FabricServiceImpl and instead
have separate lock on the properties map when manipulating or reading it.
The serverStatus flag is already volatile so nothing needs to be done for it.