Snap 1682 (#213)
Changes proposed in this pull request

A thread receives a put message, then it tries to take the lock BucketRegion.lockPrimaryStateReadLock(BucketRegion.java:1237) and waits as this is already acquired.

Another thread receives a PutMessage it takes a lock in BucketRegion.lockPrimaryStateReadLock and then tries to distribute the update to the secondary. For that it tries to get a connection and gets stuck as no connection are available.

The connections are not available because probably some threads are using them but not releasing them. 

This needs to be fixed for G. For 0.8, we are moving ahead by increasing the number of connections that can be created between two servers.

Patch testing

Ran the test on local. Running precheckin