Fixing hang in WanTest.testCreateDropGWR_Bug46171
The test checks for bind failure with an incorrect host name "bleh".
However, if such a host is resolvable (e.g. when using VPN) then the
GatewayReceiver goes into a loop assuming the bind exception is due to
port not being available and tries next port.

This change will check if the host is a local one before trying other ports.