Catalog repair procedure changes (#412)
Changes to procedure sys.repair_catalog to execute it on the server (earlier used to run on lead by sending a message to it). This will be useful to repair catalog even when lead is down. Also it now accepts two boolean params:
removeInconsistentEntries - if true then only remove the entries, otherwise just logs warnings for those
removeTablesWithData- if true will remove entries even if table contains data.