[SNAP-2568] add VTIs to public JDBC/ODBC meta-data (#426)
This set of changes is to make the Virtual Tables visible in public meta-data. There are two kinds of
VTIs: a) those that have same global view on all nodes (implementing
GfxdVTITemplateNoAllNodesRoute), b) those that are node specific and underlying engine has to
route to all nodes to get global view

The changes below add a new SYS.VTIS virtual table that lists the columns of all the inbuilt virtual
tables including itself. The type for a VTI is shown either as a "VIRTUAL TABLE" for the first case
above, or else as "LOCAL VTI" for the second case. This is then used to populate the JDBC/ODBC
getTables/getColumns/... meta-data queries.

There are other changes here to improve working with SnappyData:
- make SYS.DISKSTOREIDS as a global view so Spark layer can query it like a replicated table
- MEMBERS VTI shows type as "lead" instead of accessor with the lead server group; also mark
  currently select one as "primary lead" and remove "(normal)" suffix
- single VM case is now shown as "loner" instead of "datastore(loner)" in MEMBERS VTI

## Changes proposed in this pull request

- new SYS.VTIS table that lists all the VTIs (including itself); this is a denormalized
  table having schema/table repeated for all the columns in the underlying VTI
- use the above to supplant JDBC/ODBC getTables/getColumns APIs with full information
  about all the VTIs
- explicitly get data from all nodes in SYS.DISKSTOREIDS using a Function rather than
  marking the VTI to be sent to all nodes from outside; this is so that Spark side
  can use the VTI cleanly
- SYS.MEMBERS already has ability to get data from all nodes using a message but also
  turn that to always true for the case of Spark API calls (when "internal-connection" is true);
  this change and above allow these VTIs to appear like replicated tables and can be
  used as is by Spark API
- allow for "describe sys.members" which was failing due to MEMBERS being a keyword in
  ij shell (for SHOW MEMBERS)
- fixing a few scalastyle errors including cases where "dataTypeConversion" method was being
  invoked but the returned exception was not been thrown
- changed SYS.MEMBERS to explicitly mark "primary lead" and "lead" as VMKIND instead
  of accessor+IMPLICIT_LEADER_GROUP; also removed the "(normal)" suffixes in VMKind
  with snappy since that will always be the case in SnappyData mode
- increased default hydra master heap to 512M to fix occasional OOMEs

Other changes and fixes:

- fixing occasional failures in ShutdownDUnit: ensure that self shutdown happens only after receiving
  all the replies also added a default timeout for wait on replies (300s) rather than infinite
- cleaned up SysPoliciesVTI to just loop through obtained policies and not deal with individual
  columns; changed TABLESCHEMANAME column to SCHEMANAME
- reduced default DDL lock timeout to 3mins from 5mins
- improved SYS.MEMBERS display to show "loner" (instead of "datastore(loner)"),
  increase sizes of some columns

Fixed lock timeout issues seen in Snappy precheckin

Reason was that catalog inconsistency detection will create a connection that marked the
connection ID as CHILD_NOT_CACHEABLE. This caused no connection to be tracked on remote
nodes so GfxdDDLFinishMessage would skip any processing.

This change marks the connection as a proper one with UINITIALIZED connection ID that
will get assigned with a unique value in the EmbedConnection constructor.
Also added few logging changes to track such cases better.